move amp accordion bento component to src/bento
