+++ extensions/amp-youtube/1.0/storybook/Basic.js
@@ -5,7 +5,7 @@ import {
   BentoAccordionContent,
   BentoAccordionHeader,
   BentoAccordionSection,
-} from '#bento/components/accordion/1.0/component';
+} from '#bento/components/bento-accordion/1.0/component';
 
 import * as Preact from '#preact';
 import {useRef, useState} from '#preact';
