--- src/bento/components/accordion/OWNERS
@@ -1,14 +1,0 @@
-// For an explanation of the OWNERS rules and syntax, see:
-// https://github.com/ampproject/amp-github-apps/blob/main/owners/OWNERS.example
-
-{
-  rules: [
-    {
-      owners: [{name: 'ampproject/wg-components'}],
-    },
-    {
-      pattern: 'validator-*.protoascii',
-      owners: [{name: 'ampproject/wg-amp4email', notify: true}],
-    },
-  ],
-}
